requirements.txt
coverage
datetime
docopts
pycodestyle
pytest
pytest-cov
pytest-mock
python-dateutil
pytz
requests
requests_mock
uuid
vcrpy
google-api-python-client
google-auth-httplib2
google-auth-oauthlib